HUB-4946 Reviewer workspace - Submission inbox: Missing avatar for assignees on project overview
