+++ app/frontend/components/domains/jurisdictions/submission-inbox/application-inbox-table.tsx
@@ -16,20 +16,13 @@ import {
 } from "@chakra-ui/react"
 import { Info, Swap } from "@phosphor-icons/react"
 import { observer } from "mobx-react-lite"
-import React, { useMemo } from "react"
+import React from "react"
 import { useTranslation } from "react-i18next"
 import { Link, useNavigate } from "react-router-dom"
 import { IPermitApplication } from "../../../../models/permit-application"
 import { IPermitProject } from "../../../../models/permit-project"
-import { IUser } from "../../../../models/user"
-import { useMst } from "../../../../setup/root"
 import { IPermitApplicationInboxStore } from "../../../../stores/submission-inbox-store"
-import {
-  ECollaborationType,
-  EInboxViewMode,
-  EPermitApplicationInboxSortFields,
-  EPermitApplicationStatus,
-} from "../../../../types/enums"
+import { EInboxViewMode, EPermitApplicationInboxSortFields, EPermitApplicationStatus } from "../../../../types/enums"
 import { ISort } from "../../../../types/types"
 import { Paginator } from "../../../shared/base/inputs/paginator"
 import { PerPageSelect } from "../../../shared/base/inputs/per-page-select"
@@ -40,17 +33,14 @@ import { SearchGridItem } from "../../../shared/grid/search-grid-item"
 import { SearchGridRow } from "../../../shared/grid/search-grid-row"
 import { PermitApplicationStatusTag } from "../../../shared/permit-applications/permit-application-status-tag"
 import { SortIcon } from "../../../shared/sort-icon"
-import { DesignatedCollaboratorAssignmentPopover } from "../../permit-application/collaborator-management/designated-collaborator-assignment-popover"
+import { ApplicationReviewAssigneesCell } from "./application-review-assignees-cell"
 import { InboxNoMatchingEmpty } from "./inbox-no-matching-empty"
-import { renderAssignPlusIconTrigger, ReviewAssigneesRow } from "./review-assignees-row"
 import { SubmissionInboxMarkUnreadIconButton } from "./submission-inbox-mark-unread-icon-button"
 
 interface IProps {
   searchStore: IPermitApplicationInboxStore | IPermitProject
   applications: IPermitApplication[]
 }
-
-const MAX_VISIBLE_BLOCK_LEVEL_REVIEW_ASSIGNEE_AVATARS = 3
 
 export const ApplicationInboxTable = observer(function ApplicationInboxTable({ searchStore, applications }: IProps) {
   const { t } = useTranslation()
@@ -316,7 +306,7 @@ const ApplicationInboxRow = observer(function ApplicationInboxRow({
           e.stopPropagation()
         }}
       >
-        <ApplicationBlockLevelAndDesignatedAssigneesCell application={application} />
+        <ApplicationReviewAssigneesCell application={application} />
       </SearchGridItem>
 
       <SearchGridItem>
@@ -341,48 +331,6 @@ const ApplicationInboxRow = observer(function ApplicationInboxRow({
     </SearchGridRow>
   )
 })
-
-const ApplicationBlockLevelAndDesignatedAssigneesCell = observer(
-  function ApplicationBlockLevelAndDesignatedAssigneesCell({ application }: { application: IPermitApplication }) {
-    const { t } = useTranslation()
-    const { permitApplicationStore } = useMst()
-
-    const primaryAssignee = application.designatedReviewer?.collaborator?.user ?? null
-
-    const secondaryAssignees = useMemo(() => {
-      const seenUserIds = new Set<string>(primaryAssignee ? [primaryAssignee.id] : [])
-      const users: IUser[] = []
-
-      application.getCollaborationAssignees(ECollaborationType.review).forEach((collaboration) => {
-        const user = collaboration.collaborator?.user
-        if (user && !seenUserIds.has(user.id)) {
-          seenUserIds.add(user.id)
-          users.push(user)
-        }
-      })
-
-      return users
-    }, [application, primaryAssignee])
-
-    return (
-      <ReviewAssigneesRow
-        primaryAssignee={primaryAssignee}
-        secondaryAssignees={secondaryAssignees}
-        maxSecondaryVisible={MAX_VISIBLE_BLOCK_LEVEL_REVIEW_ASSIGNEE_AVATARS}
-        emptyText={t("ui.unassigned")}
-      >
-        <DesignatedCollaboratorAssignmentPopover
-          permitApplication={application}
-          collaborationType={ECollaborationType.review}
-          onBeforeOpen={async () => {
-            await permitApplicationStore.fetchPermitApplication(application.id, true)
-          }}
-          renderTrigger={renderAssignPlusIconTrigger({ ariaLabel: t("permitCollaboration.sidebar.title") })}
-        />
-      </ReviewAssigneesRow>
-    )
-  }
-)
 
 const ApplicationActionsMenu = observer(function ApplicationActionsMenu({
   application,
